Refactor Docker setup: update DynamoDB command to in-memory, adjust Dockerfile to copy all source files, and modify simulator logging. Add .gitignore for polygon_sp.json. Enhance order service logic to handle courier status updates more robustly and improve tracking service error handling.

+++ services/core-api/src/models.py
@@ -1,5 +1,5 @@
 import os
-from sqlalchemy import Column, Integer, String, Numeric, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, Numeric, ForeignKey
 from sqlalchemy.orm import DeclarativeBase, sessionmaker
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 from pydantic import BaseModel, ConfigDict, Field
@@ -36,7 +36,7 @@ class Restaurant(Base):
     name = Column(String(128), nullable=False)
     lat = Column(Numeric(10, 8), nullable=False)
     lon = Column(Numeric(11, 8), nullable=False)
-    h3_index = Column("h3_index", Integer, nullable=False)
+    h3_index = Column("h3_index", BigInteger, nullable=False)
     id_cuisine_type = Column("id_cuisine_type", Integer, ForeignKey("cuisinetypes.id_cuisine_type"), nullable=False)
 
 class Courier(Base):
